Move `lib` files
