+++ android/src/main/java/com/swmansion/gesturehandler/lib/Extensions.kt
@@ -1,11 +1,11 @@
-package com.swmansion.gesturehandler
+package com.swmansion.gesturehandler.lib
 
 import com.facebook.react.bridge.ReactContext
 import com.facebook.react.modules.core.DeviceEventManagerModule
 import com.facebook.react.uimanager.UIManagerModule
 
 val ReactContext.deviceEventEmitter: DeviceEventManagerModule.RCTDeviceEventEmitter
-    get() = this.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+  get() = this.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
 
 val ReactContext.UIManager: UIManagerModule
-    get() = this.getNativeModule(UIManagerModule::class.java)!!
+  get() = this.getNativeModule(UIManagerModule::class.java)!!
