+++ android/src/main/java/com/swmansion/gesturehandler/lib/GestureHandlerInteractionController.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler
+package com.swmansion.gesturehandler.lib
 
 interface GestureHandlerInteractionController {
   fun shouldWaitForHandlerFailure(handler: GestureHandler<*>, otherHandler: GestureHandler<*>): Boolean
