+++ android/src/main/java/com/swmansion/gesturehandler/lib/LongPressGestureHandler.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler
+package com.swmansion.gesturehandler.lib
 
 import android.content.Context
 import android.os.Handler
