+++ android/src/main/java/com/swmansion/gesturehandler/lib/PointerEventsConfig.kt
@@ -1,4 +1,4 @@
-package com.swmansion.gesturehandler
+package com.swmansion.gesturehandler.lib
 
 enum class PointerEventsConfig {
   /**
